% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_histogram.R
\name{create_histogram}
\alias{create_histogram}
\title{create_histogram}
\usage{
create_histogram(df, x, group, x_lab = "", y_lab = "", font_size = 10.5)
}
\arguments{
\item{df}{A training data frame or data frame extension (e.g. a tibble).}

\item{x}{The column within data_frame that should serve as the x axis.}

\item{group}{The column within data_frame that the histogram should be coloured by.}

\item{x_lab}{A string if a label for the x-axis is required. Default is "" (no axis label).}

\item{y_lab}{A string if a label for the y-axis is required. Default is "" (no axis label).}

\item{font_size}{Size of the font used in the histogram. Default is 10.5.}
}
\value{
A histogram created by ggplot.
The histogram should have different colours based on the quality of wine.
The x-label should always show, while the y-axis is only shown if add_y_lab = TRUE.
Returned histogram should work well with cowplot so the histograms are joined as one
visualization.
}
\description{
Creates ggplot histograms with x labels, different coloured qualities, and
specifiable x-axis, y-axis, labels bundled into a workflow. It is then applied
to a workflow where the input is training data, and the workflow is used
alongside cowplot commands to create the visualization.
}
\examples{
create_histogram(df = mtcars, x = mpg, group = am,
                 x_lab = "Miles/(US) gallon", y_lab = "Count", font_size = 10.5)

}
