% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_linear_reg.R
\name{fit_linear_reg}
\alias{fit_linear_reg}
\title{Fit Linear Regression Model}
\usage{
fit_linear_reg(formula, data_frame)
}
\arguments{
\item{formula}{An R regression formula.}

\item{data_frame}{A training data frame or data frame extension (e.g. a tibble).}
}
\value{
A fitted linear regression model object.
Returned object contains finalized recipe and fitted model objects (e.g.,
coefficients) inside.
}
\description{
Specifies a linear regression model, creates a recipe, and bundles them into
a workflow. Applies workflow to training data to return a fitted model.
}
\examples{
fit_linear_reg(quality ~ ., training)
}
