% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_w_new_col.R
\name{load_w_new_col}
\alias{load_w_new_col}
\title{Load dataset with new column names}
\usage{
load_w_new_col(dataset_path, col_names, delimiter)
}
\arguments{
\item{dataset_path}{An URL or path that leads to a dataset.}

\item{col_names}{Vector with the new column names as strings (quoted).}

\item{delimiter}{The dataset's delimiter (could be ".", ";", ",", etc), quoted.}
}
\value{
Loaded dataset with the new column names, as a tibble.
}
\description{
Given an URL of an online dataset, the datasets' delimiters,
and a vector of new column names, loads the dataset as a tibble
with the new column names.
}
\examples{
load_w_new_col("https://archive.ics.uci.edu/ml/machine-learning-databases/wine-quality/winequality-red.csv",
c("fixed_acidity", "volatile_acidity", "citric_acid",
"residual_sugar", "chlorides", "free_sulfur_dioxide",
"total_sulfur_dioxide", "density", "pH", "sulphates",
"alcohol", "quality"), ";")
}
